requirements.txt
google-genai
uvicorn[standard]
fastapi[standard]
openai-whisper
